+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## [5.4.2] - 2025-04-10
+
+* fix: Repair asset image detection logic
+* new feature: dio api generate hurl code
+* Update: dart assets check level error to warning
+
 ## [5.4.1] - 2025-04-08
 
 * fix some bugs